backend/pyproject.toml
[project]
name = "smart-summarizer"
version = "0.1.0"
description = ""
authors = [
    {name = "abbasi07",email = "engineer2107@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[tool.poetry]
packages = [{include = "smart_summarizer", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

